+++ CHANGELOG.md
@@ -6,13 +6,17 @@
 
 ### New features
 
+### Bugfixes
+
+## 0.8.3
+
+### New features
+
 - Added `\Wizaplace\Catalog\Declination::isBrandNew`
 - Added `\Wizaplace\Catalog\Product::getVideo`
 - Added `\Wizaplace\SDK\User\User::getBirthday`
 - Added `\Wizaplace\SDK\User\UpdateUserCommand::setBirthday`
 - Added `\Wizaplace\SDK\Catalog\ProductSummary::getShortDescription`
-
-### Bugfixes
 
 ## 0.8.2